reduce memory usage during data generation
random.normal() generates as float64 and gets converted to a smaller
dtype. generating all the needed data in a single call thus uses a lot
more memory than is necessary.

this changes it such that each frame gets generated separately.
data clipping is also changed to be in-place.

the gaussian filtering which gave the video a washed-out look is also
removed.