Use packaging library for version comparison
The release of matplotlib 3.9.1.post1 broke some of the version
comparison code that is used, as it does tuple(map(int, ...)), and in
this case `.post1` did not map to an int.  Since pytest uses the
packaging library as a dependency, we can rely on that library in our
own test suite, so we use that instead.