GraphicsLayout: Fix removeItem() regression
This fixes a regression introduced in
54efcbd2398c0f39a11ec1d14fe8f7f3f3a6b901, where the stretch factors for
the row/column previously containing a removed item were zeroed. In
general, this is of course incorrect, as there will be other items
sharing the same row/column that should not be disturbed. The intention
behind that part of the change is not clear to me; this commit simply
reverts it.